requirements.txt
flask==3.0.3
ultralytics==8.4.19
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4